feat: deprecate UserPublicKey and get rid of un-necessary seed files
