+++ apps/platform/prisma/seed.ts
@@ -1,19 +1,8 @@
 import { PrismaClient } from "@prisma/client";
 import colors from "colors";
-import seedAccesses from "./seeds/access";
-import seedBranches from "./seeds/branches";
-import seedE2EESetup from "./seeds/e2eeSetup";
-import seedInactiveUsers from "./seeds/inactiveUsers";
-import seedInvites from "./seeds/invites";
-import seedProjects from "./seeds/projects";
-import seedPullRequests from "./seeds/pullRequests";
-import seedSecrets from "./seeds/secrets";
 import seedUsers from "./seeds/users";
 
-const { Confirm } = require("enquirer");
-
 const prisma = new PrismaClient();
-
 colors.enable();
 
 if (process.env.NODE_ENV === "production") {
@@ -22,14 +11,6 @@ if (process.env.NODE_ENV === "production") {
 }
 
 const seed = async () => {
-  // const prompt = new Confirm({
-  //   name: "question",
-  //   message:
-  //     "⚠️  You are about to delete all the records on your database, and re-seed with new records, do you want to continue?"
-  //       .yellow,
-  // });
-
-  // const answer = await prompt.run();
   const answer = true;
 
   if (answer) {
@@ -38,14 +19,6 @@ const seed = async () => {
     console.log("Seeding database".underline.cyan);
     return prisma.$transaction(async () => {
       await seedUsers(10);
-      // await seedProjects(5);
-      // await seedAccesses();
-      // await seedE2EESetup();
-      // await seedBranches();
-      // await seedInactiveUsers(5);
-      // await seedInvites(5);
-      // await seedPullRequests();
-      // await seedSecrets();
     });
   } else {
     throw new Error("Seeding aborted");
@@ -61,7 +34,7 @@ const nuke = async () => {
     await prisma.branch.deleteMany();
     await prisma.projectInvite.deleteMany();
     await prisma.pullRequest.deleteMany();
-    await prisma.userPublicKey.deleteMany();
+    await prisma.keychain.deleteMany();
     await prisma.encryptedProjectKey.deleteMany();
     await prisma.secret.deleteMany();
   });
